Put options into Transaction classes
