+++ Transbank/WebpayRest/WebpayPlus/MallTransaction.cs
@@ -9,9 +9,42 @@ namespace Transbank.Webpay.WebpayPlus
 {
     public static class MallTransaction
     {
+        private static string _commerceCode = "597055555535";
+        private static string _apiKey = "579B532A7440BB0C9079DED94D31EA1615BACEB56610332264630D42D0A36B1C";
+        private static WebpayIntegrationType _integrationType = WebpayIntegrationType.Test;
+
+        public static string CommerceCode
+        {
+            get => _commerceCode;
+            set => _commerceCode = value ?? throw new ArgumentNullException(
+                nameof(value), "Commerce code can't be null."
+            );
+        }
+
+        public static string ApiKey
+        {
+            get => _apiKey;
+            set => _apiKey = value ?? throw new ArgumentNullException(
+                nameof(value), "Api Key can't be null."
+            );
+        }
+
+        public static WebpayIntegrationType IntegrationType
+        {
+            get => _integrationType;
+            set => _integrationType = value ?? throw new ArgumentNullException(
+                nameof(value), "Integration type can't be null."
+            );
+        }
+
+        public static Options DefaultOptions()
+        {
+            return new Options(CommerceCode, ApiKey, IntegrationType);
+        }
+
         public static MallCreateResponse Create(string buyOrder, string sessionId, string returnUrl, List<TransactionDetail> transactions)
         {
-            return Create(buyOrder, sessionId, returnUrl, transactions, WebpayPlus.DefaultOptions());
+            return Create(buyOrder, sessionId, returnUrl, transactions, DefaultOptions());
         }
 
         public static MallCreateResponse Create(string buyOrder, string sessionId, string returnUrl, List<TransactionDetail> transactions, Options options)
@@ -24,7 +57,7 @@ namespace Transbank.Webpay.WebpayPlus
 
         public static MallCommitResponse Commit(string token)
         {
-            return Commit(token, WebpayPlus.DefaultOptions());
+            return Commit(token, DefaultOptions());
         }
 
         public static MallCommitResponse Commit(string token, Options options)
